+++ PlantUmlViewer/PlantUmlViewer/Forms/AboutWindow.cs
@@ -12,80 +12,15 @@ namespace PlantUmlViewer.Forms
 {
     internal partial class AboutWindow : Form
     {
-        #region Assembly Attribute Accessors
-        public string AssemblyTitle
-        {
-            get
-            {
-                object[] attributes = Assembly.GetExecutingAssembly().GetCustomAttributes(typeof(AssemblyTitleAttribute), false);
-                if (attributes.Length > 0)
-                {
-                    AssemblyTitleAttribute titleAttribute = (AssemblyTitleAttribute)attributes[0];
-                    if (titleAttribute.Title != "")
-                    {
-                        return titleAttribute.Title;
-                    }
-                }
-                return System.IO.Path.GetFileNameWithoutExtension(Assembly.GetExecutingAssembly().CodeBase);
-            }
-        }
-
-        public string AssemblyVersion
-        {
-            get
-            {
-                return Assembly.GetExecutingAssembly().GetName().Version.ToString();
-            }
-        }
-
-        public string AssemblyProduct
-        {
-            get
-            {
-                object[] attributes = Assembly.GetExecutingAssembly().GetCustomAttributes(typeof(AssemblyProductAttribute), false);
-                if (attributes.Length == 0)
-                {
-                    return "";
-                }
-                return ((AssemblyProductAttribute)attributes[0]).Product;
-            }
-        }
-
-        public string AssemblyCopyright
-        {
-            get
-            {
-                object[] attributes = Assembly.GetExecutingAssembly().GetCustomAttributes(typeof(AssemblyCopyrightAttribute), false);
-                if (attributes.Length == 0)
-                {
-                    return "";
-                }
-                return ((AssemblyCopyrightAttribute)attributes[0]).Copyright;
-            }
-        }
-
-        public string AssemblyCompany
-        {
-            get
-            {
-                object[] attributes = Assembly.GetExecutingAssembly().GetCustomAttributes(typeof(AssemblyCompanyAttribute), false);
-                if (attributes.Length == 0)
-                {
-                    return "";
-                }
-                return ((AssemblyCompanyAttribute)attributes[0]).Company;
-            }
-        }
-        #endregion Assembly Attribute Accessors
-
         public AboutWindow()
         {
             InitializeComponent();
 
             this.Text = "About " + PlantUmlViewer.PLUGIN_NAME;
-            this.label_ProductName.Text = AssemblyProduct;
-            this.label_Version.Text = string.Format("Version {0}", AssemblyVersion);
-            this.label_Copyright.Text = AssemblyCopyright;
+            this.label_ProductName.Text = AssemblyAttributes.Product;
+            this.label_PluginVersion.Text = string.Format("Plugin version {0}", AssemblyAttributes.Version);
+            this.label_PlantUmlVersion.Text = string.Format("Plant UML version {0}", PlantUmlViewer.PLANT_UML_VERSION);
+            this.label_Copyright.Text = AssemblyAttributes.Copyright;
             this.richTextBox_Text.SelectedRtf = Properties.Resources.AboutText;
         }
 
@@ -139,7 +74,7 @@ namespace PlantUmlViewer.Forms
                     string latestReleaseVersion = latestRelease.Value<string>("name");
                     string latestReleaseUrl = latestRelease.Value<string>("html_url");
 
-                    if (latestReleaseVersion == AssemblyVersion)
+                    if (latestReleaseVersion == AssemblyAttributes.Version)
                     {
                         MessageBox.Show(IsDisposed ? null : this, "You are using the latest release", "Congrats", MessageBoxButtons.OK, MessageBoxIcon.Information);
                     }