+++ PlantUmlViewer/PlantUmlViewer/Forms/OptionsWindow.cs
@@ -20,6 +20,7 @@ namespace PlantUmlViewer.Forms
             textBox_JavaPath.Text = settings.Settings.JavaPath;
             textBox_Include.Text = settings.Settings.Include;
             numericUpDown_ExportSizeFactor.Value = settings.Settings.ExportSizeFactor;
+            checkBox_ExportDocument.Checked = settings.Settings.ExportDocument;
         }
 
         private void Button_Ok_Click(object sender, EventArgs e)
@@ -27,6 +28,7 @@ namespace PlantUmlViewer.Forms
             settings.Settings.JavaPath = textBox_JavaPath.Text;
             settings.Settings.Include = textBox_Include.Text;
             settings.Settings.ExportSizeFactor = numericUpDown_ExportSizeFactor.Value;
+            settings.Settings.ExportDocument = checkBox_ExportDocument.Checked;
             settings.Save();
 
             DialogResult = DialogResult.OK;