+++ PlantUmlViewer/PlantUmlViewer/Settings/PlantUmlViewerSettings.cs
@@ -5,5 +5,6 @@
         public string JavaPath { get; set; } = "";
         public string Include { get; set; } = "";
         public decimal ExportSizeFactor { get; set; } = 1;
+        public bool ExportDocument { get; set; } = true;
     }
 }