+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 - Added fluent and dark style icons
 - Updated to common zoom and scroll behaviour Ctrl + mouse wheel to zoom and (Shift +) mouse wheel to scroll vertically and horizontally
 - Added include option
+- Added metadata with option to include generated document to SVG export
 - Fixed potential crash if generation failed
 - Fixed setting Java path in options not applied without restarting Notepad++
 - Switched order of Ok and Cancel button in options window
+- Added information about used Plant UML version in about window
 - SVG.NET library updated to 3.4.4
 
 # 1.4.0.8 (07.10.2022)